feat(ui): 保存状态从底部状态栏移到顶部 toolbar 右侧
